requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
http-ece==1.1.0
idna==3.3
py-vapid==1.8.2
pycparser==2.21
pytz==2021.3
pywebpush==1.14.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
celery~=5.2.1